requirements.txt
flask==0.9
mysql-connector-python==8.2.0